Fist refactoring

+++ movies-back/src/main/java/br/uece/eesdevops/introducaospringboot/domain/exception/BookLendingNotFoundException.java
@@ -6,7 +6,7 @@ import org.springframework.web.bind.annotation.ResponseStatus;
 @ResponseStatus(HttpStatus.NOT_FOUND)
 public class BookLendingNotFoundException extends RuntimeException {
 
-    public BookLendingNotFoundException(Integer id) {
+    public BookLendingNotFoundException(Long id) {
         super("Book lending for ID " + id + " does not exist.");
     }
